backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# AI & LangChain (Updated for Python 3.14 compatibility)
openai>=2.15.0
langchain>=1.2.6
langchain-openai>=1.1.7
langchain-community>=0.4.1

# Database
sqlalchemy>=2.0.45

# Databricks (optional - for production)
databricks-sql-connector>=2.9.0

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
